Move Factor package back under algorithms.factored
