+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/factors/Semiring.scala
