+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/factors/factory/SelectFactory.scala
